4. ready

+++ controllers/blogs.js
@@ -1,11 +1,13 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', (request, response) => {
   //console.log('getting all')
   Blog
     .find({})
+    .populate('user', {username:1, name:1})
     .then(blogs => {
       response.json(blogs.map(Blog.format))
     }).catch(error => {
@@ -23,17 +25,18 @@ blogsRouter.get('/:id',(request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  if (!request.body.title ||!request.body.url) {
-    response.status(400).json({error:"Bad request"})
-    return
-  }
-  // TO DO
-  const users = await User.find({})
-  const user = users[0]
-  //console.log(user)
   const body = request.body
-  // FIX
+  const token = request.token
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!token || !decodedToken.id) {
+      return response.status(401).json({error: 'token missing or invalid'})
+    }
+    if (!request.body.title ||!request.body.url) {
+      response.status(400).json({error:"Bad request"})
+      return
+    }
+    const user = await User.findById(decodedToken.id)
     const blog = new Blog({
       title: body.title,
       author: body.author,
@@ -44,15 +47,17 @@ blogsRouter.post('/', async (request, response) => {
     const savedBlog = await blog.save()
     //console.log(savedBlog._id)
     user.blogs = user.blogs.concat(savedBlog._id)
-    console.log(User.format(user))
+    //console.log(User.format(user))
     await user.save()
     response.status(201).json(Blog.format(savedBlog))
   } catch (exception) {
-    console.log(exception)
-    response.status(500).json({error: 'Something went wrong...'})
+    if(exception.name==='JsonWebTokenError') {
+      response.status(401).json({error: exception.message})
+    } else {
+      console.log(exception)
+      response.status(500).json({error: 'Something went wrong...'})
+    }
   }
-
-
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -71,7 +76,7 @@ blogsRouter.put('/:id', async (request, response) => {
   try {
     let result = await Blog
         .findByIdAndUpdate(request.params.id, blog, {new: true})
-    console.log(Blog.format(result))
+    //console.log(Blog.format(result))
     response.status(201).json(Blog.format(result))
   } catch( exception ) {
     //console.log(exception)
@@ -80,12 +85,26 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id',  async (request, response) => {
-  console.log('deleting id ', request.params.id)
+  const token = request.token
+
+  //console.log('deleting id ', request.params.id)
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    const userId = blog.user
+    //console.log('userID: ', userId)
+
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    //console.log(decodedToken.id)
+    //console.log(userId == decodedToken.id)
+    if( userId == decodedToken.id) {
+      //console.log("in")
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).end()
+    }
   } catch (error) {
-    //console.log(error)
+    console.log(error)
     response.status(400).send({error: 'malformatted id '})
   }
 })
